Data.Array.Accelerate.LLVM.PTX

Description

This module implements a backend for the Accelerate language targeting NVPTX for execution on NVIDIA GPUs. Expressions are on-line translated into LLVM code, which is just-in-time executed in parallel on the GPU.

Documentation

data Acc a :: * -> *

Array-valued collective computations

class (Typeable * a, Typeable * (ArrRepr a)) => Arrays a

Minimal complete definition

arrays, flavour, toArr, fromArr

Synchronous execution

run :: Arrays a => Acc a -> a

Compile and run a complete embedded array program.

Note that it is recommended that you use run1 whenever possible.

runWith :: Arrays a => PTX -> Acc a -> a

As run, but execute using the specified target rather than using the default, automatically selected device.

Contexts passed to this function may all target to the same device, or to separate devices of differing compute capabilities.

run1 :: (Arrays a, Arrays b) => (Acc a -> Acc b) -> a -> b

Prepare and execute an embedded array program of one argument.

This function can be used to improve performance in cases where the array program is constant between invocations, because it enables us to bypass front-end conversion stages and move directly to the execution phase. If you have a computation applied repeatedly to different input data, use this, specifying any changing aspects of the computation via the input parameter. If the function is only evaluated once, this is equivalent to run.

To use run1 effectively you must express your program as a function of one argument. If your program takes more than one argument, you can use lift and unlift to tuple up the arguments.

At an example, once your program is expressed as a function of one argument, instead of the usual:

step :: Acc (Vector a) -> Acc (Vector b)
step = ...

simulate :: Vector a -> Vector b
simulate xs = run $ step (use xs)

Instead write:

simulate xs = run1 step xs

You can use the debugging options to check whether this is working successfully by, for example, observing no output from the -ddump-cc flag at the second and subsequent invocations.

See the programs in the 'accelerate-examples' package for examples.

run1With :: (Arrays a, Arrays b) => PTX -> (Acc a -> Acc b) -> a -> b

As run1, but execute using the specified target rather than using the default, automatically selected device.

stream :: (Arrays a, Arrays b) => (Acc a -> Acc b) -> [a] -> [b]

Stream a lazily read list of input arrays through the given program, collecting results as we go.

streamWith :: (Arrays a, Arrays b) => PTX -> (Acc a -> Acc b) -> [a] -> [b]

As stream, but execute using the specified target.

Asynchronous execution

data Async a :: * -> *

wait :: Async a -> IO a

Block the calling thread until the computation completes, then return the result.

poll :: Async a -> IO (Maybe a)

Test whether the asynchronous computation has already completed. If so, return the result, else Nothing.

cancel :: Async a -> IO ()

Cancel a running asynchronous computation.

runAsync :: Arrays a => Acc a -> IO (Async a)

As run, but run the computation asynchronously and return immediately without waiting for the result. The status of the computation can be queried using wait, poll, and cancel.

Note that a CUDA context can be active on only one host thread at a time. If you want to execute multiple computations in parallel, on the same or different devices, use runAsyncWith.

runAsyncWith :: Arrays a => PTX -> Acc a -> IO (Async a)

As runWith, but execute asynchronously. Be sure not to destroy the context, or attempt to attach it to a different host thread, before all outstanding operations have completed.

run1Async :: (Arrays a, Arrays b) => (Acc a -> Acc b) -> a -> IO (Async b)

As run1, but the computation is executed asynchronously.

run1AsyncWith :: (Arrays a, Arrays b) => PTX -> (Acc a -> Acc b) -> a -> IO (Async b)

As run1With, but execute asynchronously.

Execution targets

data PTX

The PTX execution target for NVIDIA GPUs.

The execution target carries state specific for the current execution context. The data here --- device memory and execution streams --- are implicitly tied to this CUDA execution context.

Don't store anything here that is independent of the context, for example state related to [persistent] kernel caching should _not_ go here.

createTargetForDevice :: Device -> DeviceProperties -> [ContextFlag] -> IO PTX

Create a new PTX execution target for the given device

createTargetFromContext :: Context -> IO PTX

Create a PTX execute target for the given device context

Controlling host-side allocation

registerPinnedAllocator :: IO ()

Configure the default execution target to allocate all future host-side arrays using (CUDA) pinned memory. Any newly allocated arrays will be page-locked and directly accessible from the device, enabling high-speed (asynchronous) DMA.

Note that since the amount of available pageable memory will be reduced, overall system performance can suffer.

registerPinnedAllocatorWith :: PTX -> IO ()

As with registerPinnedAllocator, but configure the given execution context.