Data.Array.Accelerate.LLVM.PTX.Foreign

Description

Documentation

data ForeignAcc f where

Constructors

ForeignAcc :: String -> (Stream -> a -> LLVM PTX b) -> ForeignAcc (a -> b)

data ForeignExp f where

Constructors

ForeignExp :: String -> IRFun1 PTX () (x -> y) -> ForeignExp (x -> y)

data LLVM target a :: * -> * -> *

The LLVM monad, for executing array computations. This consists of a stack for the LLVM execution context as well as the per-execution target specific state target.

data PTX

The PTX execution target for NVIDIA GPUs.

The execution target carries state specific for the current execution context. The data here --- device memory and execution streams --- are implicitly tied to this CUDA execution context.

Don't store anything here that is independent of the context, for example state related to [persistent] kernel caching should _not_ go here.

liftIO :: MonadIO m => forall a. IO a -> m a

Lift a computation from the IO monad.

withDevicePtr :: (ArrayElt e, ArrayPtrs e ~ Ptr a, Typeable e, Typeable a, Storable a) => ArrayData e -> (DevicePtr a -> LLVM PTX (Maybe Event, r)) -> LLVM PTX r

Lookup the device memory associated with a given host array and do something with it.

module Data.Array.Accelerate.LLVM.Array.Data

copyToHostLazy :: Arrays arrs => arrs -> LLVM PTX arrs

Copy an array from the remote device to the host. Although the Accelerate program is hyper-strict and will evaluate the computation as soon as any part of it is demanded, the individual array payloads are copied back to the host _only_ as they are demanded by the Haskell program. This has several consequences:

If the device has multiple memcpy engines, only one will be used. The transfers are however associated with a non-default stream.

Using seq to force an Array to head-normal form will initiate the computation, but not transfer the results back to the host. Requesting an array element or using deepseq to force to normal form is required to actually transfer the data.

type Async a = AsyncR PTX a

type Stream = Lifetime Stream

type Event = Lifetime Event

module Data.Array.Accelerate.LLVM.Execute.Async