Data.Array.Accelerate.LLVM.PTX.Foreign

Description

Documentation

data ForeignAcc f where

Constructors

ForeignAcc :: String -> (Stream -> a -> LLVM PTX b) -> ForeignAcc (a -> b)

data ForeignExp f where

Constructors

ForeignExp :: String -> IRFun1 PTX () (x -> y) -> ForeignExp (x -> y)

data LLVM target a :: * -> * -> *

The LLVM monad, for executing array computations. This consists of a stack for the LLVM execution context as well as the per-execution target specific state target.

data PTX

The PTX execution target for NVIDIA GPUs.

The execution target carries state specific for the current execution context. The data here --- device memory and execution streams --- are implicitly tied to this CUDA execution context.

Don't store anything here that is independent of the context, for example state related to [persistent] kernel caching should _not_ go here.

Constructors

PTX

Fields

ptxContext :: !Context

ptxMemoryTable :: !MemoryTable

ptxKernelTable :: !KernelTable

ptxStreamReservoir :: !Reservoir

fillP :: !Executable

data Context

An execution context, which is tied to a specific device and CUDA execution context.

Constructors

Context

Fields

deviceProperties :: !DeviceProperties

deviceContext :: !(Lifetime Context)

liftIO :: MonadIO m => forall a. IO a -> m a

Lift a computation from the IO monad.

withDevicePtr :: (ArrayElt e, ArrayPtrs e ~ Ptr a, Typeable e, Typeable a, Storable a) => ArrayData e -> (DevicePtr a -> LLVM PTX (Maybe Event, r)) -> LLVM PTX r

Lookup the device memory associated with a given host array and do something with it.

module Data.Array.Accelerate.LLVM.Array.Data

copyToHostLazy :: Arrays arrs => arrs -> LLVM PTX arrs

Copy an array from the remote device to the host. Although the Accelerate program is hyper-strict and will evaluate the computation as soon as any part of it is demanded, the individual array payloads are copied back to the host _only_ as they are demanded by the Haskell program. This has several consequences:

If the device has multiple memcpy engines, only one will be used. The transfers are however associated with a non-default stream.

Using seq to force an Array to head-normal form will initiate the computation, but not transfer the results back to the host. Requesting an array element or using deepseq to force to normal form is required to actually transfer the data.

cloneArrayAsync :: (Shape sh, Elt e) => Stream -> Array sh e -> LLVM PTX (Array sh e)

Clone an array into a newly allocated array on the device.

type Async a = AsyncR PTX a

type Stream = Lifetime Stream

A Stream represents an independent sequence of computations executed on the GPU. Operations in different streams may be executed concurrently with each other, but operations in the same stream can never overlap. Events can be used for efficient cross-stream synchronisation.

type Event = Lifetime Event

Events can be used for efficient device-side synchronisation between execution streams and between the host.

module Data.Array.Accelerate.LLVM.Execute.Async

Orphan instances

Foreign PTX

Methods

foreignAcc :: (Foreign asm, Typeable * a, Typeable * b) => PTX -> asm (a -> b) -> Maybe (StreamR PTX -> a -> LLVM PTX b)

foreignExp :: (Foreign asm, Typeable * x, Typeable * y) => PTX -> asm (x -> y) -> Maybe (IRFun1 PTX () (x -> y))